dataflow/streaming_opencv/requirements.txt
numpy
google-cloud-storage
